Replaced byte with int

+++ Lab5/src/ca/mcgill/ecse211/color/CanColor.java
@@ -1,13 +1,13 @@
 package ca.mcgill.ecse211.color;
 //TODO: add stddev
 public enum CanColor {
-  RED("Red", new byte[] {47, 18, 16}), GREEN("Green", new byte[] {18, 46, 20}), BLUE("Blue",
-      new byte[] {24, 43, 57}), YELLOW("Yellow", new byte[] {61, 43, 31}), UNKOWN;
+  RED("Red", new int[] {47, 18, 16}), GREEN("Green", new int[] {18, 46, 20}), BLUE("Blue",
+      new int[] {24, 43, 57}), YELLOW("Yellow", new int[] {61, 43, 31}), UNKOWN;
 
   private String name;
-  private byte[] avgRGB;
+  private int[] avgRGB;
 
-  CanColor(String name, byte[] c) {
+  CanColor(String name, int[] c) {
     this.name = name;
     avgRGB = c;
   }
@@ -37,7 +37,7 @@ public enum CanColor {
    * @param c A byte array of r g and b values
    * @return the color w/ the closest euclidean distance to c
    */
-  public static CanColor getClosestColor(byte[] color) {
+  public static CanColor getClosestColor(int[] color) {
     CanColor closest = UNKOWN;
     double minDist = 28;
     for (CanColor c : new CanColor[] {RED, GREEN, YELLOW, BLUE}) {
@@ -49,7 +49,7 @@ public enum CanColor {
     return closest;
   }
 
-  public double distanceTo(byte[] c) {
+  public double distanceTo(int[] c) {
     int sum = 0;
     for (int i = 0; i < c.length; i++) {
       sum += Math.pow((int) c[i] - avgRGB[i], 2);
